skin_lesion_classifier/requirements.txt
numpy>=1.19.2
tensorflow>=2.4.0
opencv-python>=4.5.1
scikit-learn>=0.24.1
imbalanced-learn>=0.8.0
pyyaml>=5.4.1
matplotlib>=3.3.4
seaborn>=0.11.1